[prep CDF-25018] 📋 Support set-pending-ids for files (#1710)
# Description

**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is step 3 of that process, enabling linking of asset-centric files
to cognite-files, such that the new CogniteFile can point to an existing
asset-centric file and thus not needing to download and reupload the
file content.

In this PR, I introduce `.set_pending_ids` for files which is set on
asset-centric files in preparation for the creation of a CogniteFile.
**Reviewer**: This follows the same pattern that we are doing for
timeseries.

## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip