[CDF-25194] 👷‍♂️ Update instance source API (#1735)
# Description

Context: This is part of the alpha migration tooling in Toolkit, for
overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is step 4 of that process, enabling downloading the instance source
of migrated resources. This will, for example, be used by the Canvas
migration to lookup the instance Id of asset-centric resources in a
Chart.

Note the `InstanceSource` classes are generated by `pygen`.

![image](https://github.com/user-attachments/assets/39339156-9f03-4ddf-92b0-7b2c8be76e87)


## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip