[CDF-25286] 🦾 Improve producer worker (#1769)
# Description

**Context**: The `ProducerWorker` is an internal structure that
implements a producer/worker pattern for doing tasks concurrently.

In addition, found a bug in the `ProduceWorker` that did not handle if
the iter call fails. This does not impact the end-user as the
`ProducerWorker` always has a valid iterable when called from the CLI
commands.

## Changelog

- [x] Patch
- [ ] Minor
- [ ] Skip

## cdf

### Improved

- When running `cdf dump asset/timeseries`, Toolkit have fixed the the
process bar descriptions.

## templates

No changes.