[CDF-25192] 👮 Respect frequency of Raw Row count (#1739)
# Description

**Context**: We have a `cdf profile raw` command that counts the number
of rows in raw tables. This is currently using the `/profile/raw`
endpoint, however, this only works for up to 10,000 rows. We plan to use
the function `raw_row_count` which uses `run/query` as a fallback. To
avoid overloading transformations we agreed with the transformation team
that it will only count up to 500,000 rows and can not be invoked by the
user more than every 10 minutes.

This PR implements the limit of 10 minutes. I achieve this by storing
when the `raw_row_count` function was called in a file, and if it called
again within 10 minutes it raises a `RunTimeError`. Note that we allow
to call the `raw_row_count` function multiple times in CLI command, but
the 10 minutes limit is between commands.

See #1727 for usage


## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip