[CDF-25255, CDF-25197] 🐣 API Source for view classes (#1761)
# Description

**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is step 1 of that process.

This PR adds a API support for the ViewSource class (we already have it
for the Instance Source). With API support, I mean that we extend the
internal ToolkitClient that is an extension of CogniteClient with CRUD
methods for reading and writing ViewSource objects.


## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip