[CDF-25285] 🦿Convert asset-centric properties to container properties Part 1 (#1777)
# Description

**Reviewer** This is a ~250 lines of code, with more than 300 lines of
tests.

**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is step 3 of that process.

This method will be used to convert the asset-centric
properties/metadata to the appropriate container property.

This is part 1 of likely 3 parts. I am splitting this up to try to limit
the size. Current plan:
* Part 1 (this PR): Conversion of non-list values.
* Part 2 Conversion of list-values
* Part 3 Special cases from asset-centric to data modeling. For example,
going from timeseries to `CogniteTimeSeries` the property `isString`
must convert from a boolean to an enum.

## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip