[CDF-25298] 🔨Validate access utility (#1774)
# Description

This utility class will be used in commands to inform/stop the user due
to lack of authorization.

For example, the plan is to use it in the `cdf purge command`. 
```python
validator = ValidateAccess(client, 'purge')
if space_ids := validator.instances(["read", "write"]):
    self.warn(LimitedAccessWarning(f"You can only purge instances in the {len(instance_scope.space_ids)} spaces you have access to: {humanize_collection(space_ids)}.")

if timeseries_access := validtor.instance(["read", "write"], operation="unlink"):
     scopes = [f"{scope_name} ({humanize_collection(ids)})" for scope_name, ids in timeseries_access]
     self.warn(LimitedAccessWarning(f"You can only unlink time series in the following scopes: {humanize_collection(scopes)}."
```


## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip