[CDF-25334] 👹 Storage IO interface (#1807)
# Description

**Context**: This introduces a new interface into Toolkit. The intention
of this interface is to use it for downloading and uploading data. This
will be used in the `cdf migrate/download/upload/purge` commands. I
expect to implement a class for raw, all asset-centric resources, and
instances.

The motivation for having a separate PR for this interface, is to limit
the size of the PRs for this new functionality. Note that there is no
tests in this PR as this is just an interface and thus there is nothing
to test.

## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip