[CDF-25256] 😁 Add ViewSourceLoader (#1767)
# Description

<del>Stacked on #1761 </del>

**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is step 1 of that process.

This PR adds ViewSourceLoader. The ResourceLoader is a standardized
interface we use in Toolkit to do CRUD operations. This enables use to
support storing mapping between asset-centric resources and views. The
intention is that in a migration process, the migration team develops
this mapping in a dev environment, then, they are stored and pushed to
the prod (and staging) environment, such that they can be use do execute
the migration there in the exact same way as it was done in the dev
environment.

## Changelog

- [x] Patch
- [ ] Minor
- [ ] Skip

## cdf

### Added

- [alpha] Added support for storing mapping between asset-centric
resources and views in configurations with the new resource type
`ViewSource`.

## templates

No changes.