[CDF-25201] ⏩ HTTP processor (#1737)
# Description

**Reviewer**: This code change is +400 lines in the main code base,
while the remaining lines are tests.

**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is step 3 of that process were we need to write a lot of nodes to
DMS.

In Cognite Toolkit, we have several commands (migrate/purge) that do a
lot of requests in which a few of them may fail. Currently, we use the
`cognite-sdk` to do this requests, but this has several weaknesses:

* **It is not robust against poison items**. The PySDK will do fail
first and stop doing requests when doing large uploads. We instead want
to continue and instead report on the items that are failing.
* **Speed**. The PySDK limits the number of workers you can use for data
modeling to 1. When in practice the concurrency limits are higher. This
is problematic in large jobs such as migration and purging data, as it
causes them to be a magnitude slower than necessary.
* <del>**Rate limiting**. In the Cognite-SDK if you hit multiple 429,
only the current worker stops to do a request. Not all workers.</del>
* **Not rely on cognite-sdk internals** Today, Toolkit support multiple
alpha and pre-alpha endpoints which are not yet in the SDK. To do that
we have relied on internals of the PySDK, which we have been critiqued
by riks-reviewers earlier. This enables us to have our own logic doing
requests, and thus alpha and pre-alpha endpoints can be supported in
Toolkit without using internals of the PySDK.

This PR introduces a `HTTPBatchProcessor`. It is designed to handle all
the weaknesses above, and is intended to be used in the migrate
commands. It will also be used in the current purge command, and planned
commands download/upload.

In addition, we sometimes need to implement new alpha endpoints that are
not yet available in the `cognite-sdk`, today we rely on private methods
in the SDK. This is risky as pointed out [earlier
risk-reviews](https://github.com/cognitedata/toolkit/pull/1666#discussion_r2174988387),
this new `HTTPBatchProcessor` will enable us to not use the private
methods and classes of the `cognite-sdk`.


## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip