[prep CDF-25523] 🐛Update cognite sdk (#1840)
# Description

For this alpha command we depend on some internals in the `cognite-sdk`,
which had a breaking change. **Note** We are in the process on replacing
this dependency of the internal of `cognite-sdk`, but as a stop-gap
solution this PR fixes the issue this time.

## Changelog

- [x] Patch
- [ ] Skip

## cdf

### Fixed

- [alpha] The commands `cdf migrate files/timeseries` no longer fails
with `TypeError` when upgrading the `cognite-sdk` to >=`7.78.1`.

## templates

No changes.