[CDF-25319] 🫠 Introduce batch processor (#1784)
# Description

This introduces a new version of the `HTTPProcessor`, see
https://github.com/cognitedata/toolkit/pull/1737 for argument why I
introduce this one.

The existing version of the `HTTPProcessor` takes in an itreable and
uses rich to create a progress bar. This version of the `HTTPProcessor`
allows you more granular control over adding items. I need this in an
upcoming PR.



## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip