[CDF-25523] 🎱Batching in datapoints subscriptions (#1841)
# Description

The datapoint subscription supports up to 10,000 timeseries in one
subscription. However, you cannot include this in a single request when
you create it. Thus, you have to done one create request and up to 99
update requests with a maximum of 100 timeseries per request. This PR
does that split when creating datapoints subscriptions. **Note** The 100
and 10,000 limit is the sum of the instance and timeseries ID count.

Note there <del>is</del> was a bug in the `cognite-sdk`, the
`description` parameter is not updated. This have to go in first:
https://github.com/cognitedata/cognite-sdk-python/pull/2255. [UPDATE]
Merged Ready.

## Changelog

- [x] Patch
- [ ] Skip

## cdf

### Fixed

- When deploying datapoints subscriptions and specifying timeseries IDs,
Toolkit now correctly splits request such that you can have up to 10,000
timeseries IDs in one subscription. Before you could only have the
request limit which is 100 timeseries IDs.

## templates

No changes.