[CDF-25294] 🙃 Output file migration (#1782)
# Description

**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is step 3 of that process.

This PR introduces a result output file. This is a `csv` file that
contain one line for each resource that is migrated with the results of
that resources.

To set up the writing to file, I needed the response status of each
individual item, so I used the new `HTTPProcessor` and combined it with
the `ProducerWorker`.

## Changelog

- [ ] Patch
- [ ] Minor
- [x] Skip