[CDF-24985] 📈 Charts API (#1844)
# Description

**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is step 4 of that process.

This PR introduces a client implementation for the Charts API. We need
this to migrate charts such that we can download charts with
asset-centric timeseries/events and migrate them to Cognite
timeseries/activities.

## Changelog

- [ ] Patch
- [x] Skip