[CDF-25604]🏗 Introducing migration issues (#1869)
# Description

**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is step 3 of that process.

This PR introduces DTO classes for issues that can occur during a
migration. This will be written to file when the user runs a migration
command such as `cdf migrate timeseries`.

## Changelog

- [ ] Patch
- [x] Skip