[CDF-24664] 🐛 Modules pull failing (#1894)
# Description

**Reviewer** The fix is +12 lines, while the test are ~250 lines.

**Context**: The pull command allows the user to pull down configuration
from the fusion UI to local YAML configurations. This is useful when a
governed resource, for example, a transformation is updated in the UI
and you want the updated version in your local configuration while
maintaining the location on disk and built variables.

This PR fixes a bug in which the selection of resources to pull fails
when the Toolkit skips resources due to mix of relative and absolute
filepaths. To fix this bug, I introduces a test battery for the
selection logic and the fixed the bug.

## Changelog

- [x] Patch
- [ ] Skip

## cdf

### Fixed

- The command `cdf modules pull` now correctly selecting resources given
as both relative and absolute paths.

## templates

No changes.