[CDF-25595] 🐛 Set nonce function schedules (#1891)
# Description

Instead of letting the `cognite-sdk` set the credentials, we create the
session first and set the `nonce`. This ensures that function schedules
follows the same pattern we use for transformations.

This requires updating the `cognite-sdk` which leads to a few tests
needs update, and found the issue with workflow trigger described below.

## Changelog

- [x] Patch
- [ ] Skip

## cdf

### Improved

- When deploying a function schedule, if the credentials are invalid,
Toolkit now gives an improved error message.
- When deploying a function schedule, Toolkit now has a small
performance boost.

### Fixed

- Toolkit will no longer redeploy a workflow trigger without timeszone
specified that is unchanged.

## templates

No changes.