[CDF-24820] ⛏Added workflow version YAML (#1898)
# Description

We are in the progress of adding pydantic classes to match the YAML
format Toolkit expects for configurations. The goal is to give better
error message to the user on syntax errors.

This PR introduces the WorkflowVersionYAML class to validate workflow
triggers. Note this is not yet exposed, so no change to the end user.

## Changelog

- [ ] Patch
- [x] Skip