[prep CDF-25549] 👷 Prepare asset download (#1881)
# Description

This is preparation for introducing a download/upload/migrate asset
command.

We have an interface in Toolkit for downloading/uploading data.
`StorageIO`. To select data we use a selector. In this PR, I change the
selector for asset-centric data to one of three subtypes. This makes for
a cleaner implementation.



## Changelog

- [ ] Patch
- [x] Skip