[CDF-25371] Instance IO implementation (download part) (#1875)
# Description

Stacked on #1876

**Context** StorageIO is an interface for downloading and uploading data
in Toolkit. This is intended to be used with multiple commands
download/upload/purge/migration.

This PR implements the download part of the interface for instances.
This is because this is all we need right now for the purge command. And
additional benefit is that this PR becomes very light to review.

## Changelog

- [ ] Patch
- [x] Skip