[CDF-25373] 🗂 Filemetadata IO (#1901)
# Description


This is preparation for introducing a download/upload/migrate files
command.

We have an interface in Toolkit for downloading/uploading data.
`StorageIO`. To select data we use a selector. In this PR, I add an
implementation of this interface for FileMetadata. In addition, I
refactored the AssetIO such that I could reuse most of it for
FileMetadat.

## Changelog

- [ ] Patch
- [x] Skip