[CDF-24803] 🥠 Exraction pipeline config yaml (#1910)
# Description

We are in the progress of adding pydantic classes to match the YAML
format Toolkit expects for configurations. The goal is to give better
error message to the user on syntax errors.

This PR introduces the ExtractionPipelineConfigYAML class to validate
extraction pipeline config. Note this is not yet exposed, so no change
to the end user.

## Changelog

- [ ] Patch
- [x] Skip