[CDF-24816] Transformation notification yaml (#1908)
# Description

We are in the progress of adding pydantic classes to match the YAML
format Toolkit expects for configurations. The goal is to give better
error message to the user on syntax errors.

This PR introduces the TransformationNotificationYAML class to validate
transformation notifications. Note this is not yet exposed, so no change
to the end user.


## Changelog

- [ ] Patch
- [x] Skip