[CDF-24808] 🎈 Relationship yaml (#1935)
# Description

We are in the progress of adding pydantic classes to match the YAML
format Toolkit expects for configurations. The goal is to give better
error message to the user on syntax errors.

This PR introduces the RelatoinshipYAMl to validate relationships. Note
this is not yet exposed, so no change to the end user.

## Changelog

- [ ] Patch
- [x] Skip