[CDF-24807] 🎍 Hosted Extractor Mapping YAML (#1933)
# Description

We are in the progress of adding pydantic classes to match the YAML
format Toolkit expects for configurations. The goal is to give better
error message to the user on syntax errors.

This PR introduces the HostedExtractorMappingYAMLclass to validate
hosted extractor mappings. Note this is not yet exposed, so no change to
the end user.

## Changelog

- [ ] Patch
- [x] Skip