[CDF-24798] Pydantic class for data model (#1914)
# Description

We are in the progress of adding pydantic classes to match the YAML
format Toolkit expects for configurations. The goal is to give better
error message to the user on syntax errors.

This PR introduces the `DataModelYAML` class to validate data model.
This PR also contains `CONTAINER_AND_VIEW_EXTERNAL_ID_PATTERN ` and
`VIEW_VERSION_PATTERN` constant renaming to generic names
`EXTERNAL_ID_PATTERN` and `VERSION_PATTERN` respectively as these
patterns are applicable for all `externalId` and `version`.

## Changelog

- [ ] Patch
- [x] Skip

## cdf

### Added

- `DataModelYAML` class for data model validation

## templates

No changes.