[CDF-25698] 🤓 HTTP Client Part 2 Items (#1923)
# Description

<del>Stacked on #1069 <del>

**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is step 3 of that process.

This PR extends the HTTPClient with special handling of Item requests.
This will be used in migration to ensure the command is idempotent. For
example, when you migrate timeseries, you link to the existing. If you
try to link an existing node, then the API will give a 400. This ensures
that if there is a single bad item in a request of 1000 item, all the
other 999 will go through.


## Changelog

- [ ] Patch
- [x] Skip