[CDF-24809] Pydantic class for Robotic capabilities (#1956)
# Description

We are in the progress of adding pydantic classes to match the YAML
format Toolkit expects for configurations. The goal is to give better
error message to the user on syntax errors.

This PR introduces the `RobotCapabilityYAML` class to validate robotics
capabilities resource. this is the base branch for all the robotic
resource validation models that will be added.

## Changelog

- [ ] Patch
- [x] Skip

## cdf

### Added

- `RobotCapabilityYAML` class for robotics capabilities validation.

## templates

No changes.