[CDF-24804] 😀 Hosted extractors sources YAML (#1947)
# Description


We are in the progress of adding pydantic classes to match the YAML
format Toolkit expects for configurations. The goal is to give better
error message to the user on syntax errors.

This PR introduces the HostedExtractorSourceYAMLclass to validate hosted
extractor source. Note this is not yet exposed, so no change to the end
user.

## Changelog

- [ ] Patch
- [x] Skip