[CDF-25753] 👷‍♂️ Rework purge spaces part 1 (#1977)
# Description

This is part 1 of reimplementing the `cdf purge space` commands. It only
includes the `dry-run` mode and is not connected to any command. The
motivation for only doing the dry-run is to keep this PR small and fast
to review.

The motivation for this reimplementation is as follows:
* The `/data-modeling/statistics` endpoint makes dry-run a lot faster to
implement, but requires a new approach.
* We have new components such as HTTPClient and WorkerProducer that will
be very useful for running the purge code.
* The current implementation lacks proper testing, a reimplementation
will ensure it is well tested.

## Changelog

- [ ] Patch
- [x] Skip