[CDF-24801] Pydantic class for Edge (#1949)
# Description

We are in the progress of adding pydantic classes to match the YAML
format Toolkit expects for configurations. The goal is to give better
error message to the user on syntax errors.

This PR introduces the `EdgeYAML` class to validate edges. This PR is
stacked on #1948 because edge uses `NodeType` to specify start node and
end node of the edge.

## Changelog

- [ ] Patch
- [x] Skip

## cdf

### Added

- `EdgeYAML` class for edge validation

## templates

No changes.