[CDF-25765] 🪲 Parsing error (#1982)
# Description

An edge case bug where we complain about invalid YAML when the config is
valid YAML. In addition, I found two more issues: The wanning message
failed to set the resource ID and YAML should be a mapping and not array

## Changelog

- [x] Patch
- [ ] Skip

## cdf

### Improved

- When deploying an extractor pipeline config, Toolkit now gives an
improved error message if the config property is not YAML and no longer
gives an incorrect warning if a `!keyvalut` token is used with multiple
preceding spaces.

## templates

No changes.