[CDF-25777] Made checksum a high-priority warning instead of throwing an error (#1985)
# Description

Users can fetch modules (deployment packs) from external libraries. This
is an alpha feature.

They specify the library url and checksum in their cdf toml. The
checksum is there to verify the integrity of the package.

```
[alpha_flags]
external_libraries = true

[library.foo]
url = "https://github.com/cognitedata/foo/package.zip"
checksum = "sha256:barbarbarbarbarbarbar"
```

Packages may be updated unbeknownst to the user, making the checksum
invalid. This used to throw an **error**, blocking the users from
completing their task and frustrating them.

This change prints a warning instead of throwing an error, letting the
user decide if they want to complete download and review it, or if they
want to break and verify the checksum with the source.

## Changelog

- [x] Patch
- [ ] Skip

## cdf

### Changed

- Warn if checksum doesn't match, but let user continue

## templates

No changes.