[CDF-25801] GraphQL YAML (#2000)
# Description

We are in the progress of adding pydantic classes to match the YAML
format Toolkit expects for configurations. The goal is to give better
error message to the user on syntax errors.

This PR introduces the `GraphQLDataModelYAML` to validate graphql data
model configurations. Note this is not yet exposed, so no change to the
end user. Note this is a hidden alpha feature only used by two
customers, but to complete the transformation to use pydantic for
validation of configurations we also need to cover this.

## Changelog

- [ ] Patch
- [x] Skip