Skip tests (#2013)
# Description

**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).

We are in the process of the migration model, and these two tests
depends on the model being deployed in the the Project.

We need to skip them for now, but they can be added after #2004 has been
merged.

## Changelog

- [ ] Patch
- [x] Skip