[CDF-25783, CDF-25803]  🗺 Added resource mapping yaml (#1988)
# Description

We are in the progress of adding pydantic classes to match the YAML
format Toolkit expects for configurations. The goal is to give better
error message to the user on syntax errors.

This PR introduces the ResourceViewMappingYAML to validate mappings from
asset-centric to views. This will be used as part of the migration.

## Changelog

- [ ] Patch
- [x] Skip