[prep CDF-25832] 🚩Look-up DM only flag (#2010)
# Description

We need to support different behavior depending on whether it is a
DM-only project or Hybrid.

This PR implements the backend call to check what type of project
Toolkit is connected to using the method described here:
https://cognitedata.slack.com/archives/C043Z2ZV9KN/p1724078664012789

In addition, I updated our `HTTPClient` with the following

* The `request_with_retries` now return a `ResponseList` that have a
`raise_for_status` and `get_first_body` helper.
* Fixed circular imports for `HTTPClient` and `ToolkitConfig`.


## Changelog

- [ ] Patch
- [x] Skip