[CDF-25833] 😁 Rework space part3 (#2011)
# Description

This is part 3 of reimplementing the `cdf purge space` commands. It
makes the default behavior to skip nodes with datapoints and file
content, instead of deleting.

The motivation for this reimplementation is as follows:
* The `/data-modeling/statistics` endpoint makes dry-run a lot faster to
implement, but requires a new approach.
* We have new components such as HTTPClient and WorkerProducer that will
be very useful for running the purge code.
* The current implementation lacks proper testing, a reimplementation
will ensure it is well tested.


## Changelog

- [ ] Patch
- [x] Skip