[CDF-25769] 🤩 Added validation to migration command. (#2022)
# Description


**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is step 3 and4 of that process.

In Toolkit, a command class contains the logic that is executed when a
CLI command is called.

This PR is a copy-paste of the validation logic from the
`BaseMigrateCommand` to `MigrationCommand`. Currently, there is a one
migration command for each resource type the Toolkit supports. We are in
the process of replacing all of these with a single command,
`MigrationCommand`. that contains the logic for migrating any resource
and uses a implementations of a shared interface for each resource type.
Once the `MigrationCommand` is finished, I will replace all the other
migration commands.

The tests are also copy-pasted over. This is such that when all the old
migration commands are replaced, the old tests can be deleted.

## Changelog

- [ ] Patch
- [x] Skip